push

+++ src/db/schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE "users"(
-    "id" BIGINT NOT NULL,
+    "id" SERIAL NOT NULL,
     "email" VARCHAR(255) NOT NULL,
     "password" VARCHAR(255) NOT NULL
 );
